+++ examples/dom/table.cpp
@@ -54,6 +54,10 @@ int main() {
   content.DecorateCellsAlternateRow(color(Color::Cyan), 3, 1);
   content.DecorateCellsAlternateRow(color(Color::White), 3, 2);
 
+  // Decorate 2 random cells with a red border.
+  table.SelectCell(3, 4).Border(LIGHT, color(Color::Red));
+  table.SelectCell(2, 7).Border(LIGHT, color(Color::Red));
+
   auto document = table.Render();
   auto screen =
       Screen::Create(Dimension::Fit(document, /*extend_beyond_screen=*/true));